src/main/java/ax/ha/tdd/chess/engine/pieces/Queen.java
package ax.ha.tdd.chess.engine.pieces;

import ax.ha.tdd.chess.engine.Chessboard;
import ax.ha.tdd.chess.engine.Coordinates;
import ax.ha.tdd.chess.engine.Player;

public class Queen extends ChessPiece {
    public Queen(Player player, Coordinates location) {
        super(PieceType.QUEEN, player, location);
    }

    public boolean move(Chessboard board, Coordinates destination) {
        if (!canMove(board, destination))
            return false;
        board.removePiece(this);
        location = destination;
        board.addPiece(this);
        return true;
    }

    @Override
    public String getSymbol() {
        return pieceType.getSymbol();
    }

    @Override
    public boolean canMove(Chessboard chessboard, Coordinates destination) {
        final int targetPosX = destination.getX();
        final int targetPosY = destination.getY();
        final int curretPosX = location.getX();
        final int curretPosY = location.getY();
        if (targetPosX == curretPosX || targetPosY == curretPosY) {
            return new Rook(player, location).canMove(chessboard, destination);
        } else {
            return new Bishop(player, location).canMove(chessboard, destination);
        }
    }
}